move deferred into test/
